backend/requirements.txt
fastapi
uvicorn
tensorflow=2.18.0
numpy
opencv-python
spotipy
python-multipart
python-dotenv
